services/media-processor/pyproject.toml
[project]
name = "media-processor-service"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "aiobotocore>=2.22.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "celery>=5.5.2",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "ruff>=0.11.9",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[tool.ruff]
line-length = 100
exclude = ["src/video/transcoder.py"]
